# POC: Improve database throughput.