--- a/src/batch_write.rs
+++ b/src/batch_write.rs
@@ -50,7 +50,7 @@ pub(crate) struct WriteBatchRequest {
 
 impl DbInner {
     #[allow(clippy::panic)]
-    async fn write_batch(&self, batch: WriteBatch) -> Result<Arc<KVTable>, SlateDBError> {
+    pub async fn write_batch(&self, batch: WriteBatch) -> Result<Arc<KVTable>, SlateDBError> {
         let now = self.options.clock.now();
 
         let current_table = if self.wal_enabled() {