@dmvk 2024-12-05
I've also run the following test using bencher:

```
     Running `target/release/bencher --path /slatedb-bencher_100_1 db --db-options-path /Users/david/Workspace/slatedb/slatedb/src/bencher/Slatedb.toml --duration 60 --key-generator FixedSet --key-count 100000 --val-len 256 --block-cache-size 134217728 --put-percentage 100 --concurrency 1`
```

On main:

![100_1](https://github.com/user-attachments/assets/98eeabe7-9f19-4d3c-8ae0-a5d7f3b24a2e)

With the patch:

![100_1](https://github.com/user-attachments/assets/272a515c-97bd-4eab-8b6b-80dd5c1298e4)

With the patch + https://github.com/slatedb/slatedb/pull/375


![100_1](https://github.com/user-attachments/assets/ec16128e-bcee-43b4-89f4-8a4c2a2f8e23)