@dmvk 2024-12-05
flamegraph from the last bencher run (With the patch + https://github.com/slatedb/slatedb/pull/375); this looks pretty reasonable with skiplistmap being the main bottleneck

![flamegraph](https://github.com/user-attachments/assets/f557ff52-3477-47cd-9315-be7518a4e74d)