@criccomini 2024-12-06
> I think the write through an actor model existed for a reason.

The thinking was essentially:

1. We were already locking on write, so it was essentially single threaded anyway.
2. We needed a centralized place to serialize all writes so we could assign sequence numbers when #260 is implemented.
3. Pebble [followed a similar design](https://github.com/cockroachdb/pebble/blob/master/docs/rocksdb.md#commit-pipeline) with some reasonable justification.

I'm definitely open to changing the design if it leads to the kind of performance gains we're seeing here. Though, one question I have: are these `bencher` runs going against in-memory, local filesystem, local object storage, or remote object storage? I'm a bit sensitive about chasing performance improvements at the expense of complexity of the gains are very minimal in actual production workloads (i.e. remote object storage). IIRC, bencher runs on local FS right now, though I plan to make it run against remote in the future. Do you have any numbers when using a remote object store before/after this patch?