@criccomini 2024-12-12
@dmvk Wanted to touch base with you on this one. Do you have a flame graph handy for `main` with your workload right now? I'd be interested in seeing that.

I was reviewing the Pebble docs and they outline their approach:

> 1. Lock commit mutex
> * Add batch to commit queue
> * Assign batch sequence number
> * Write batch to the WAL
> 2. Unlock commit mutex
> 3. Apply batch to memtable (concurrently)
> 4. Publish batch sequence number

This is _slightly_ better than what we do. We hold the write lock all the way through step (3) (`maybe_freeze_wal` or `maybe_freeze_memtable` in `write_batch.rs`). I'm rather curious if the flame graph shows a significant amount of time waiting on the wal/memtable freezes. We're doing `std::mem::replace` when freezing, so it's not like we're iterating over kv pairs, but it could still be comparatively slow. We could fix this, as Pebble does, by making wal/memtable freezes lockless using a sequence number to gate reads against batches that are still being copied.